Task-7: Add database to configuration
