+++ config/config.yaml
@@ -10,3 +10,11 @@ gateway:
   host: 0.0.0.0
   port: 8080
 
+database:
+  host: database
+  port: 5432
+  user: postgres
+  password: postgres
+  name: ova_journey_api
+  sslMode: disable
+  driver: pgx
