+++ internal/config/configuration_test.go
@@ -36,6 +36,15 @@ func TestConfiguration_LoadConfigurationFromFile(t *testing.T) {
 						Host: "127.0.0.1",
 						Port: 8080,
 					},
+					Database: &DatabaseConfiguration{
+						Host:     "database",
+						Port:     5432,
+						User:     "postgres",
+						Password: "postgres",
+						Name:     "ova_journey_api",
+						SslMode:  "disable",
+						Driver:   "pgx",
+					},
 				},
 				err: nil,
 			},
