+++ internal/config/project_configuration.go
@@ -1,0 +1,7 @@
+package config
+
+// ProjectConfiguration type represents configuration with information about project
+type ProjectConfiguration struct {
+	Name    string `yaml:"name"`
+	Version string `yaml:"version"`
+}
